Reduced redundant codes
